changes in employee category
